project/roomie_user/serializers.py
from rest_framework import serializers
from .models import Profile

class ProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = Profile
        fields = ['bio', 'date_of_birth', 'profile_picture', 'role', 'roomie_score']